# use the technically-correct copy()